+++ packages/zowe-explorer/src/tools/ZoweLocalStorage.ts
@@ -66,7 +66,7 @@ export class LocalStorageAccess extends ZoweLocalStorage {
         [Definitions.LocalStorageKey.SETTINGS_OLD_SETTINGS_MIGRATED]: StorageAccessLevel.Read,
         [Definitions.LocalStorageKey.ENCODING_HISTORY]: StorageAccessLevel.Read | StorageAccessLevel.Write,
         [Definitions.LocalStorageKey.DS_SEARCH_OPTIONS]: StorageAccessLevel.Read | StorageAccessLevel.Write,
-        [Definitions.LocalStorageKey.SHOW_RELEASE_NOTES_VERSION]: StorageAccessLevel.Read | StorageAccessLevel.Write,
+        [Definitions.LocalStorageKey.DISPLAY_RELEASE_NOTES_VERSION]: StorageAccessLevel.Read | StorageAccessLevel.Write,
         [PersistenceSchemaEnum.Dataset]: StorageAccessLevel.Read | StorageAccessLevel.Write,
         [PersistenceSchemaEnum.USS]: StorageAccessLevel.Read | StorageAccessLevel.Write,
         [PersistenceSchemaEnum.Job]: StorageAccessLevel.Read | StorageAccessLevel.Write,